Publishing Amartya on Github
Amartya means immortal in our sanskrit language.

This application is created to aware the masses about the freedom fighters that laid their lives for the freedom of our India.

Their contribution is a tribute towards the nation building & we as Indians have to carry that legacy to make their vision about India a reality.

Amartya implements immersive technology & artificial intelligence that will be a gateway to a whole new level of User experience.

Through this app, User will be able to interact with the objects coming from the digital world into the real world, as well as take a tour of the virtual world.

Our vision is to share knowledge in an interactive & most accessible way that will educate the users and bring a positive result on the society as a whole.